Factor coordinate system utilities out to separate file, with detailed explanation
